+++ map/map_first.cub
@@ -7,22 +7,23 @@ F 220,100,0
 C 225,30,0
 
          111111111111111111111111
-         100000100000000000000001
-  111111100000000000000000111111
-  1S0000100000000111000001
-  100000100000000000000001
-  100000100000000000000001
-  100000100000000000000001
-  100000111111111000000001
+         1000000000000000000001
+  11111111000000000000001111111
+  1S1111111111111111111111
   100000000000000000000001
-  100000000000000000000001
-  100000000000000000000001
-  100000101001000000000001
-  100000101001000000000001
-  100000101001000000000001
-  100000101001000000000001
-  100000101001000001000001
-  100000101001000001000001
-  100000000000000001000001
-  100000000000000000000001
+  100000000000000010000001
+  1000000000000001 10000001
+  100000000000000010000001
+  111111111111111111111111
+  111111111111111111111111
+
+  111111111111111111111111
+  111111111111111111111111
+  111111111111111111111111
+  111111111111111111111111
+  111111111111111111111111
+  111111111111111111111111
+  111111111111111111111111
+  111111111111111111111111
+  111111111111111111111111
   111111111111111111111111
