+++ src/main.c
@@ -6,7 +6,7 @@
 /*   By: lugonzal <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/07 20:53:48 by lugonzal          #+#    #+#             */
-/*   Updated: 2022/11/09 20:07:37 by lugonzal         ###   ########.fr       */
+/*   Updated: 2022/11/30 19:28:47 by lugonzal         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -15,22 +15,21 @@
 #include "mlx.h"
 #include "custom_errors.h"
 
-
 int	main(int argc, char **argv)
 {
 	t_application	appl;
 	t_mlx			*mlx_win;
 	t_map			*map;
-    t_custom_error  c_error;
+	t_custom_error	c_error;
 
-	ft_memset(&c_error, 0, sizeof(t_application));
+	ft_memset(&c_error, 0, sizeof(t_custom_error));
 	ft_memset(&appl, 0, sizeof(t_application));
 	custom_error_init(&c_error);
 	if (argc != 2 || application_init(&appl, argv[1], &c_error) < 0)
-    {
+	{
 		application_destory(&appl);
-        return (print_custom_error(&c_error));
-    }
+		return (print_custom_error(&c_error));
+	}
 	mlx_win = &appl.mlx_win;
 	map = &appl.map;
 	if (!map->map)
